adds browser with replacements
